Problem: redundant cardano-wallet/ in info.rkt
Launcher won't work, because it looks one level too deep, which doesn't exist.

Solution: Correct info.rkt.

Now we have the problem that:

    (sched (msg-iip fvm in (cons 'add a-graph)))

Expects the path to `(node)` to `(make-graph)` to be a string, not a path.

```
#(struct:exn:fail:contract string-append: contract violation
  expected: string?
  given: #<path:/nix/store/zdycm6sk2arwrbarw4yiwyaxz6n8lylz-fractalide-env/share/racket/pkgs/fractalide/modules/rkt/rkt-fbp/agents/test/main.rkt>
  argument position: 2nd
  other arguments...:
   "agents/"
   ".rkt" #<continuation-mark-set>)
```

But before fixing this, it's probably better to bump to latest
fractalide/fractalide, as a lot has been happening there the latest
couple of weeks.